added tracks and playlist
