+++ server.js
@@ -9,7 +9,6 @@ const publicRouter = require('./data/routers/publicRouter.js');
 const authRouter = require('./data/routers/authRouter.js');
 const registerRouter = require('./data/routers/registerRouter.js');
 const loginRouter = require('./data/routers/loginRouter.js');
-const trackRouter = require('./data/routers/trackRouter.js');
 
 const server = express();
 
@@ -29,7 +28,6 @@ server.use('/api', publicRouter);
 server.use('/api/login', loginRouter);
 server.use('/api/register', registerRouter);
 server.use('/api/auth', authenticate, authRouter);
-server.use('/api/tracks', trackRouter);
 
 // Exports
 module.exports = server;
